Added simulation clock source to testbench.
There's a waveform in the simulation now. :-)

I got this from the language templates but used my own, more descriptive, names.

Some might say the names are too descriptive because if the oscillator frequency was changed all the names would need changing. In a small example project, for a newbie, it helps to be overly descriptive.