Added PLL using IP Clocking Wizard.
This is my first attempt at using IP from the IP Catalog.

The clock is simple enough that it is not hard to use without the Clock Wizard but Clock Wizard is nice.

Vivado Design Suite User Guide Design Flows Overview UG892 says it is only necessary to add the 'The IP source file (.xci)' to the repo, so that what I've done. I'm assuming all the other files can be generated using the 'Generate Output Products...' menu.