Include 'Clock Wizard' generated clock in the project.
This should have been part of the last commit. Clearly I didn't notice the import change in the project file because there were a number of other changes.

Dealing with the xpr in the repo is a bit of pain pain because there is stuff in it that keeps changing but I don't understand why.