Added AXI GPIO IP core and joined it to the AXI traffic generator.
The AXI traffic generator wasn't doing much on its own, it needs a slave to respond before any traffic actually flows.

Adding the a AXI GPIO made it possible to seem the AXI traffic in the simulation waveform.

I had to increase 'Max Clocks to Run', before I did this the data got stuck at '2'.

AXI GPIO v2.0 LogiCORE IP Product Guide PG144 tells us that address 0 in the AXI GPIO is the GPIO data which is a R/W register. So by changing the address to 0 in address.coe we write various values to GPIO data! :-)